+++ packages/gatsby-plugin-offline/src/app-shell.js
@@ -2,7 +2,7 @@ import React from "react"
 
 class AppShell extends React.Component {
   render() {
-    return <div />
+    return <span />
   }
 }